Initial commit of an empty Angular app
Angular was installed in the root of this repo using the following
command:

  npm install @angular/cli

The Angular app was created using the following command and anwswers:

  ng new web
  ? Would you like to add Angular routing? Yes
  ? Which stylesheet format would you like to use? SCSS   [ https://sass-lang.com/documentation/syntax#scss ]

The env file can sources to add the dir where `ng` is to the PATH. It
can be sourced by doing `. ./env` from the root of the repo.