Add a GitHub Actions workflow for deploying to App Engine (#12)
This PR adds a GitHub Actions workflow that deploys to Google App Engine for
each create/update of a pull request and each push to the main branch. The PR
also includes an `app.yaml` App Engine configuration file for the deployment
(the Node.js runtime is used but only the support for static files is used).

The formats of the deployed App Engine version are `pr#-<commit hash prefix>`
for pull requests and `main-<commit hash prefix>` for pushes to the main
branch.