Fix the App Engine version for pushes to the main branch (#16)
This PR fixes the missing commit hash from the App Engine versions produced by
the GitHub Actions workflow triggered for pushes to the main branch.

The problem was that `github.event.pull_request.head.sha` only works for pull
requests. The solution for pushes to the main branch is `github.sha`.