Add support for Protocol Buffers and some sample data (#13)
This PR adds a proto definition for the data we are going to work with.
The protobuf.js [1] library is used to generate the TypeScript code from
the .proto file. The advantages of using protobuf.js instead of the
compiler provided by Google [2] is that:

  (a) it provides support for TypeScript
  (b) it comes readily available via npm
  (b) it provides a superior parsing performance.

The protobuf.js can generate static code, use a JSON representation of
the .proto files or use the .proto files directly. To keep things simple
this PR is relying on the static code. The steps for generating the code
can be run using `npm run genproto`.

This PR also adds an initial set of sample data: each .pb from the
src/assets/ contains a single PushInfos message.

[1] https://github.com/protobufjs/protobuf.js
[2] https://github.com/protocolbuffers/protobuf