Implement screenshot testing using Happo (#18)
This PR integrates screenshot testing remotely using Happo.io. Testing locally
requires comparing a served web page to a baseline screenshot. However, due to
slight differences in machines, when running the image comparison using a CI
(CircleCI, TravisCI, and Github Actions) results in an error in the test because
the images are slightly shifted.

Testing remotely using Happo circumvents this issue by taking a screenshot of
the page when deployed and comparing it to a previously deployed version.
Because the screenshots happen on the same machine (remotely using Happo), the
screenshots are able to align when they are the same. Another advantage of Happo
is its ability to visit the site on multiple browsers and display the
differences. One disadvantage is that it only compares static websites after it
is running on App Engine. Local testing is possible with Happo but requires more
set up.

A proposed pipeline is as follows (assuming we will be comparing a pull request
to main):

1. The most recent commit to main is automatically set as the golden image that
we are going to compare against.

2. When we are ready to perform the screenshot testing, we can open a PR. Using
Github Actions, it will automatically deploy the app to a static URL. Then,
Happo will automatically take screenshots of this deployed site and compare it
against the baseline images from the commit that we specified.

3. Happo will tell us if there are differences. If there are no differences,
then a check will appear on the pull request checks. If there are differences,
there will be a "X" on the Happo check, and clicking on the "details" link will
lead us to a site where we can compare the screenshots side by side and either
"accept" or "reject" the differences. After the differences are accepted or
rejected, the check on the pull request will be updated to say "Accepted by
NAME" or "Rejected by NAME".

4. If the differences are accepted and the pull request is merged, the golden hash
will be automatically updated and will be used as the golden images for the future
PRs.