Fix mistake of getting most recent commit hash to main (#22)
There was a typo in the previous PR with `git rev parse` getting the most recent
commit to main. After changing it to `git rev-parse`, the following error occurs:
`ambiguous argument 'origin/main': unknown revision or path not in the working tree.`
I am instead using `git ls-remote` to list and get the most recent commit to main.