Skip the postinstall script in production mode
The pinning we do in the postinstall script is not necessary in
production mode and is causing problems in App Engine because the Cloud
Build is running that step.

This PR also switches the App Engine runtime to nodejs12 which showed
this issue consistently. Based on the error logs from Cloud Build I
suspect the update that is going on for the nodejs10 runtime will also
run the postinstall script.