Set up local screenshot testing using protractor-image-comparison (#20)
This PR integrates screenshot testing locally using protractor image comparison.
Testing locally requires comparing a served web page to a baseline screenshot.
However, due to slight differences in machines, when running the image
comparison using a CI (CircleCI, TravisCI, and Github Actions) results in an
error in the test because the images are slightly shifted. Local screenshot
testing can still be used via the `npm run screenshot` command but will not be
used as one of the CI steps. The local screenshot testing takes a screenshot of
the served web app and compares it to a baseline screenshot. The baseline
screenshots can be reset by deleting the files and running `npm run screenshot` to
autopopulate the baseline images with the current served app. Running this
command without any baseline images will automatically result in no differences
between the baseline and currently served images, resulting in a passing test.