Add Happo checks for all-pushes and one-push pages (#27)
This PR adds screenshot testing to the all-pushes and one-push pages. Originally,
it was only testing the home page. Now, it includes testing of the all-pushes pages
with 11k and 83 pushes. It also tests the one-push page with a completed push
and a built unchanged push for the push def with 11k pushes as well as a
reverted and completed push for the push def with 83 pushes. Currently, there
is an issue with the testing of the all-pushes page because it takes a screenshot
before the table is loaded in. This will be fixed in a subsequent PR.

This PR also deletes an unnecessary TypeScript plugin and removes testing on
a Firefox browser.