Add an initial static bar chart for the one-push pages (#30)
This PR includes components of a static top bar chart and static a bottom
bar chart. It also adds a static brush selector for the bottom bar chart.
In future, a drop down menu, a boxplot, and a tooltip will be added to the
page, and the brush selector would perform zoom in and out for the graph.