Add dot plot and percentile lines to CDF plot on one-push page (#31)
This PR adds a dot plot and lines representing percentiles on top of the
existing static CDF plot. The dot plot places the dots in a non-overlapping
method that displays the quantity and raw duration of the completed pushes.
The percentile lines are placed at the 10%, 50% and 90% values, if visually
possible. If they overlap, then the 10% and 90% lines are moved outward until
there is enough space. If there is no space, then only the 50% line appears on
the graph. Additionally, there is a line that appears on the graph of the
completed pushes that represent how the current push compares to all the
other pushes.