Set the max-line-length to 80 in tslint.json (#33)
The 140 was set by default by Angular generator. The Google JavaScript asks for
a line length of 80. This PR is also installing reformatting the code using the
clang-format.

The code extension that can do that on save is xaver.clang-format. The snippet
from `settings.json` for code-server that is doing that is the following:

    {
      [...]
      "typescript.format.insertSpaceAfterOpeningAndBeforeClosingNonemptyBraces": false,
      "[typescript]": {
        "editor.defaultFormatter": "xaver.clang-format",
        "editor.formatOnSave": true
      },
      [...]
    }