Add interactive tooltip to the bar chart and fix time display (#38)
This PR adds an interactive tooltip to the bar chart, which displays the push information
when the user hovers on a bar or the empty area above it. During hovering, the bar will
be highlighted, and the point corresponding to the bar in the boxplot will also be highlighted
and outlined in black. The interactions are triggered by the `mouseover` and `mouseleave`
events.

The startTime on the x axis is also fixed by reformatting x labels to `yyyy:MM:dd HH:mm:ss`
so that bars don't present on the same x position. The missing tick on the y axis is added.
The next PR will change the duration unit depending on the time of the pushes, fix ticks and
labels overlaying for the boxplot and remove `any` in type declaration.