Add user interaction to CDF chart (#41)
This PR adds event handling for mouse clicks and mouse hovers. When a click
occurs on the graph, the area to the left of the clicked value will be shaded lighter,
the dots will change to grey and lines will appear to signify the probability and the
number of points below that value. When the mouse hovers over the graph, two
rulers will appear to show the duration and probability value that the point
corresponds to.

This PR also changes the name of the duration helper file from duration.utils.ts
to duration-utils.ts.