Add toggle to show and hide dot plot (#43)
This PR uses a material design toggle slide to allow the user to show or hide
the dot plot. The dots can be distracting and can cover the graph but provides
valuable information, thus, having the user alternate between the two is helpful.
The default is set showing the dots.