Reflow the README.md to 80-characters and remove extra spaces
In #47 I made the mistake of not wrapping the text to 80-characters (I
copy/pasted from the Google Doc we were working on). That was a mistake
that cause the review for the rest of the changes to the README.md to be
hard.

The extra spaces removed in this PR were at the end of various lines.
Ideally we whould have caught that in the review but they were probably
hard to spot in the GitHub interface.